Restructed hermes folder structure
