+++ server/hermes/gcloud/storage.go
